+++ CHANGES.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for `asContract()` on SmartWeave.
 - Support for `contractCaller()` on SmartWeave.
 - Support for `txSender()` on SmartWeave.